Move filter_version and locked decorator to templates module, AlreadyRunning to exceptions module
